requirements.txt
opencv_python==4.1.0.25
PyYAML==5.1.1
